Add `bind(c)` keyword to some zFORp types and functions, so they can be referenced from C tests

+++ tests/fortran/testFortran.f
@@ -23,7 +23,7 @@ program main
 
   ! zfp_stream
   type(zFORp_stream_type) :: zfp_stream
-  real :: desired_rate, rate_result
+  real (kind=8) :: desired_rate, rate_result
   integer :: dims, wra
   integer :: zfp_type
